+++ domain/src/main/java/com/project/personal/domain/model/WeatherDetailsModel.kt
@@ -1,15 +1,15 @@
 package com.project.personal.domain.model
 
 class WeatherDetailsModel(
-        val id: Int,
+        val id: Long,
         val applicableDate: String,
         val weatherState: String,
         val weatherStateAbbr: String,
         val windSpeed: Float,
         val windDirection: String,
-        val minTemp: Int,
-        val maxTemp: Int,
-        val currentTemp: Int,
+        val minTemp: Double,
+        val maxTemp: Double,
+        val currentTemp: Double,
         val airPressure: Float,
         val humidity: Float,
         val visibility: Float,
